+++ packages/react-native/src/icons/app/AlignTextCenter.tsx
@@ -1,0 +1,22 @@
+import Svg, { Path } from "react-native-svg";
+import type { SvgProps } from "react-native-svg";
+import { Ref, forwardRef } from "react";
+const SvgAlignTextCenter = (props: SvgProps, ref: Ref<Svg>) => (
+  <Svg
+    fill="none"
+    viewBox="0 0 24 24"
+    className={props.className}
+    ref={ref}
+    {...props}
+  >
+    <Path
+      stroke="currentColor"
+      strokeLinecap="round"
+      strokeLinejoin="round"
+      strokeWidth={1.3}
+      d="M5 8h14M5 12h14M9 16h6"
+    />
+  </Svg>
+);
+const ForwardRef = forwardRef(SvgAlignTextCenter);
+export default ForwardRef;