+++ packages/react-native/src/icons/app/CheckCircleLine.tsx
@@ -1,0 +1,23 @@
+import Svg, { Circle, Path } from "react-native-svg";
+import type { SvgProps } from "react-native-svg";
+import { Ref, forwardRef } from "react";
+const SvgCheckCircleLine = (props: SvgProps, ref: Ref<Svg>) => (
+  <Svg
+    fill="none"
+    viewBox="0 0 24 24"
+    className={props.className}
+    ref={ref}
+    {...props}
+  >
+    <Circle cx={12} cy={12} r={8} stroke="currentColor" strokeWidth={1.3} />
+    <Path
+      stroke="currentColor"
+      strokeLinecap="round"
+      strokeLinejoin="round"
+      strokeWidth={1.3}
+      d="m9 12 2.4 2.4L15 9.6"
+    />
+  </Svg>
+);
+const ForwardRef = forwardRef(SvgCheckCircleLine);
+export default ForwardRef;