+++ packages/react-native/src/icons/app/EyeVisible.tsx
@@ -1,0 +1,22 @@
+import Svg, { Path, Circle } from "react-native-svg";
+import type { SvgProps } from "react-native-svg";
+import { Ref, forwardRef } from "react";
+const SvgEyeVisible = (props: SvgProps, ref: Ref<Svg>) => (
+  <Svg
+    fill="none"
+    viewBox="0 0 24 24"
+    className={props.className}
+    ref={ref}
+    {...props}
+  >
+    <Path
+      stroke="currentColor"
+      strokeLinejoin="round"
+      strokeWidth={1.3}
+      d="M20 12c-1-3-4-6-8-6s-7 3-8 6c1 3 4 6 8 6s7-3 8-6Z"
+    />
+    <Circle cx={12} cy={12} r={2.35} stroke="currentColor" strokeWidth={1.3} />
+  </Svg>
+);
+const ForwardRef = forwardRef(SvgEyeVisible);
+export default ForwardRef;