+++ packages/react-native/src/index.ts
@@ -1,2 +1,6 @@
 // Export components
-export * from "./components/ExampleComponent"
+export * from "./components/ExampleComponent";
+export * from "./components/Icon";
+
+// Export icons
+export * from "./icons";