+++ packages/react/src/experimental/OneDataCollection/Filters/FilterTypes/InFilter/index.tsx
@@ -1,0 +1,38 @@
+import type { BaseFilterDefinition } from "../"
+import { FilterTypeDefinition } from "../types"
+import { InFilter } from "./InFilter"
+import { FilterOption, InFilterOptions } from "./types"
+import { loadOptions } from "./useLoadOptions"
+
+export const inFilter: FilterTypeDefinition<
+  string[],
+  InFilterOptions<string>
+> = {
+  emptyValue: [],
+  isEmpty: (value) => (value || []).length === 0,
+  render: (props) => <InFilter {...props} />,
+  chipLabel: async (value, { schema }) => {
+    const options = await loadOptions(schema.options)
+
+    const selectedLabels = value.map((v) => {
+      const option = options.find((opt) => opt.value === v)
+      return option?.label ?? String(v)
+    })
+
+    const firstSelectedLabel = selectedLabels[0]
+    const remainingCount = selectedLabels.length - 1
+    const hasMultipleSelections = remainingCount > 0
+
+    return hasMultipleSelections
+      ? `${firstSelectedLabel} +${remainingCount}`
+      : `${firstSelectedLabel}`
+  },
+}
+
+export default inFilter
+
+export type InFilterDefinition<T = unknown> = BaseFilterDefinition<"in"> & {
+  options:
+    | Array<FilterOption<T>>
+    | (() => Array<FilterOption<T>> | Promise<Array<FilterOption<T>>>)
+}