+++ packages/react/src/lib/providers/i18n/i18n-provider-defaults.ts
@@ -5,8 +5,11 @@ export const defaultTranslations = {
     next: "Next",
   },
   actions: {
+    add: "Add",
+    edit: "Edit",
     save: "Save",
     cancel: "Cancel",
+    copy: "Copy",
     showAll: "Show all",
     showLess: "Show less",
     skipToContent: "Skip to content",